Рефакторит папку с типами
