+++ src/types/location.ts
@@ -1,0 +1,5 @@
+export type LocationT = {
+  latitude: number;
+  longitude: number;
+  zoom: number;
+};
