+++ src/types/review.ts
@@ -1,0 +1,9 @@
+import { UserT } from './user';
+
+export type ReviewT = {
+  id: string;
+  date: string;
+  user: UserT;
+  comment: string;
+  rating: number;
+};
